[UNTESTED] DirectVobSubFilter: stop overriding known BT.601 matrices
Commit e928f83920b04bb33e0df69c74bbaa8d04a47260 allowed colorimetry
information to be received from the DirectShow filter chain, so we don’t
need to rely on resolution-based guesses as much.
However, since it only checked for BT709 (and assigned a suboptimal
fallback for SMPTE240) but not BT601, a known-to-be BT601 matrix
was still overriden by the fallback guess for larger video resolutions.